project_export_20250825_234251/code/requirements.txt
Django==4.2.7
Pillow==10.1.0
django-cors-headers==4.3.1
python-decouple==3.8
PyPDF2==3.0.1
pdf2image==1.16.3
